+++ spec/defines/container_spec.rb
@@ -293,7 +293,7 @@ describe 'podman::container' do
         is_expected.to contain_exec('podman_remove_image_testing-namevar').with(
           {
             'notify'      => 'Exec[podman_create_testing-namevar]',
-            'require'     => ['Podman::Rootless[testing]', 'Service[podman systemd-logind]', 'Exec[podman_remove_container_testing-namevar]'],
+            'require'     => ['Podman::Rootless[testing]', 'Exec[podman_remove_container_testing-namevar]'],
             'environment' => ['HOME=/home/testing', 'XDG_RUNTIME_DIR=/run/user/3333', 'DBUS_SESSION_BUS_ADDRESS=unix:path=/run/user/3333/bus'],
             'cwd'         => '/home/testing',
             'user'        => 'testing',
@@ -329,7 +329,7 @@ describe 'podman::container' do
             'command'     => 'podman generate systemd  namevar > /home/testing/.config/systemd/user/podman-namevar.service',
             'refreshonly' => true,
             'notify'      => 'Exec[service_podman_testing-namevar]',
-            'require'     => ['Podman::Rootless[testing]', 'Service[podman systemd-logind]'],
+            'require'     => ['Podman::Rootless[testing]'],
             'path'        => '/sbin:/usr/sbin:/bin:/usr/bin',
             'environment' => ['HOME=/home/testing', 'XDG_RUNTIME_DIR=/run/user/3333', 'DBUS_SESSION_BUS_ADDRESS=unix:path=/run/user/3333/bus'],
             'cwd'         => '/home/testing',
@@ -343,7 +343,7 @@ describe 'podman::container' do
           {
             'command'     => "systemctl --user  enable podman-namevar.service\nsystemctl --user  start podman-namevar.service\n",
             'unless'      => "systemctl --user  is-active podman-namevar.service &&   systemctl --user  is-enabled podman-namevar.service\n",
-            'require'     => ['Podman::Rootless[testing]', 'Service[podman systemd-logind]'],
+            'require'     => ['Podman::Rootless[testing]'],
             'path'        => '/sbin:/usr/sbin:/bin:/usr/bin',
             'environment' => ['HOME=/home/testing', 'XDG_RUNTIME_DIR=/run/user/3333', 'DBUS_SESSION_BUS_ADDRESS=unix:path=/run/user/3333/bus'],
             'cwd'         => '/home/testing',
@@ -354,8 +354,8 @@ describe 'podman::container' do
 
       # only here to reach 100% resource coverage
       context 'cover additional resource from other classes' do
-        it { is_expected.to contain_exec('loginctl_linger_testing') }            # from podman::rootless
-        it { is_expected.to contain_service('podman systemd-logind') }           # from podman::rootless
+        it { is_expected.to contain_podman__rootless('testing') }
+        it { is_expected.to contain_loginctl_user('testing') }                   # from podman::rootless
         it { is_expected.to contain_file('/home/testing/.config') }              # from podman::rootless
         it { is_expected.to contain_file('/home/testing/.config/systemd') }      # from podman::rootless
         it { is_expected.to contain_file('/home/testing/.config/systemd/user') } # from podman::rootless
@@ -415,7 +415,7 @@ describe 'podman::container' do
       it do
         is_expected.to contain_exec('podman_remove_image_testing-namevar').with(
           {
-            'require'     => ['Podman::Rootless[testing]', 'Service[podman systemd-logind]', 'Exec[podman_remove_container_testing-namevar]'],
+            'require'     => ['Podman::Rootless[testing]', 'Exec[podman_remove_container_testing-namevar]'],
             'environment' => ['HOME=/home/testing', 'XDG_RUNTIME_DIR=/run/user/3333', 'DBUS_SESSION_BUS_ADDRESS=unix:path=/run/user/3333/bus'],
             'cwd'         => '/home/testing',
             'user'        => 'testing',
@@ -426,7 +426,7 @@ describe 'podman::container' do
       it do
         is_expected.to contain_file('/home/testing/.config/systemd/user/podman-namevar.service').with(
           {
-            'require' => ['Podman::Rootless[testing]', 'Service[podman systemd-logind]', 'Exec[service_podman_testing-namevar]'],
+            'require' => ['Podman::Rootless[testing]', 'Exec[service_podman_testing-namevar]'],
           },
         )
       end
@@ -655,7 +655,7 @@ describe 'podman::container' do
             'command'     => "systemctl --user  stop podman-namevar\nsystemctl --user  disable podman-namevar\n",
             'onlyif'      => "test \"$(systemctl --user  is-active podman-namevar 2>&1)\" = \"active\" -o   \"$(systemctl --user  is-enabled podman-namevar 2>&1)\" = \"enabled\"\n",
             'notify'      => 'Exec[podman_remove_container_user-namevar]',
-            'require'     => ['Podman::Rootless[user]', 'Service[podman systemd-logind]'],
+            'require'     => ['Podman::Rootless[user]'],
             'path'        => '/sbin:/usr/sbin:/bin:/usr/bin',
             'environment' => ['HOME=/home/user', 'XDG_RUNTIME_DIR=/run/user/3333', 'DBUS_SESSION_BUS_ADDRESS=unix:path=/run/user/3333/bus'],
             'cwd'         => '/home/user',
@@ -670,7 +670,7 @@ describe 'podman::container' do
             'command'     => 'podman container rm --force namevar',
             'unless'      => 'podman container exists namevar; test $? -eq 1',
             'notify'      => 'Exec[podman_remove_image_user-namevar]',
-            'require'     => ['Podman::Rootless[user]', 'Service[podman systemd-logind]'],
+            'require'     => ['Podman::Rootless[user]'],
             'path'        => '/sbin:/usr/sbin:/bin:/usr/bin',
             'environment' => ['HOME=/home/user', 'XDG_RUNTIME_DIR=/run/user/3333', 'DBUS_SESSION_BUS_ADDRESS=unix:path=/run/user/3333/bus'],
             'cwd'         => '/home/user',
@@ -685,7 +685,7 @@ describe 'podman::container' do
             'provider'    => 'shell',
             'command'     => 'podman rmi mandatory:latest || exit 0',
             'refreshonly' => true,
-            'require'     => ['Podman::Rootless[user]', 'Service[podman systemd-logind]', 'Exec[podman_remove_container_user-namevar]'],
+            'require'     => ['Podman::Rootless[user]', 'Exec[podman_remove_container_user-namevar]'],
             'path'        => '/sbin:/usr/sbin:/bin:/usr/bin',
             'environment' => ['HOME=/home/user', 'XDG_RUNTIME_DIR=/run/user/3333', 'DBUS_SESSION_BUS_ADDRESS=unix:path=/run/user/3333/bus'],
             'cwd'         => '/home/user',
@@ -698,7 +698,7 @@ describe 'podman::container' do
         is_expected.to contain_file('/home/user/.config/systemd/user/podman-namevar.service').only_with(
           {
             'ensure'  => 'absent',
-            'require' => ['Podman::Rootless[user]', 'Service[podman systemd-logind]', 'Exec[service_podman_user-namevar]'],
+            'require' => ['Podman::Rootless[user]', 'Exec[service_podman_user-namevar]'],
             'notify'  => 'Exec[podman_systemd_user_reload]',
           },
         )
@@ -706,7 +706,8 @@ describe 'podman::container' do
 
       # only here to reach 100% resource coverage]
       context 'cover additional resource from other classes' do
-        it { is_expected.to contain_exec('loginctl_linger_user') }            # from podman::rootless
+        it { is_expected.to contain_podman__rootless('user') }
+        it { is_expected.to contain_loginctl_user('user') }                   # from podman::rootless
         it { is_expected.to contain_file('/home/user/.config') }              # from podman::rootless
         it { is_expected.to contain_file('/home/user/.config/systemd') }      # from podman::rootless
         it { is_expected.to contain_file('/home/user/.config/systemd/user') } # from podman::rootless