+++ spec/defines/image_spec.rb
@@ -101,13 +101,14 @@ describe 'podman::image' do
             'cwd'         => '/home/dummy',
             'provider'    => 'shell',
             'user'        => 'dummy',
-            'require'     => ['Podman::Rootless[dummy]', 'Service[podman systemd-logind]'],
+            'require'     => ['Podman::Rootless[dummy]'],
           },
         )
       end
 
       # only here to reach 100% resource coverage
-      it { is_expected.to contain_exec('loginctl_linger_dummy') }            # from podman::rootless
+      it { is_expected.to contain_podman__rootless('dummy') }
+      it { is_expected.to contain_loginctl_user('dummy') }                   # from podman::rootless
       it { is_expected.to contain_exec('start_dummy.slice') }                # from podman::rootless
       it { is_expected.to contain_file('/home/dummy/.config') }              # from podman::rootless
       it { is_expected.to contain_file('/home/dummy/.config/systemd') }      # from podman::rootless
@@ -141,7 +142,7 @@ describe 'podman::image' do
             'cwd'         => '/home/dummy',
             'provider'    => 'shell',
             'user'        => 'dummy',
-            'require'     => ['Podman::Rootless[dummy]', 'Service[podman systemd-logind]'],
+            'require'     => ['Podman::Rootless[dummy]'],
           },
         )
       end
@@ -198,18 +199,18 @@ describe 'podman::image' do
             'cwd'         => '/home/testing',
             'provider'    => 'shell',
             'user'        => 'testing',
-            'require'     => ['Podman::Rootless[testing]', 'Service[podman systemd-logind]'],
+            'require'     => ['Podman::Rootless[testing]'],
           },
         )
       end
 
       # only here to reach 100% resource coverage
-      it { is_expected.to contain_exec('loginctl_linger_testing') }            # from podman::rootless
+      it { is_expected.to contain_podman__rootless('testing') }
+      it { is_expected.to contain_loginctl_user('testing') }                   # from podman::rootless
       it { is_expected.to contain_exec('start_testing.slice') }                # from podman::rootless
       it { is_expected.to contain_file('/home/testing/.config') }              # from podman::rootless
       it { is_expected.to contain_file('/home/testing/.config/systemd') }      # from podman::rootless
       it { is_expected.to contain_file('/home/testing/.config/systemd/user') } # from podman::rootless
-      it { is_expected.to contain_service('podman systemd-logind') }           # from podman::rootless
     end
 
     context 'with user set to valid testing when ensure is set to valid absent' do
@@ -239,7 +240,7 @@ describe 'podman::image' do
             'cwd'         => '/home/testing',
             'provider'    => 'shell',
             'user'        => 'testing',
-            'require'     => ['Podman::Rootless[testing]', 'Service[podman systemd-logind]'],
+            'require'     => ['Podman::Rootless[testing]'],
           },
         )
       end